+++ apps/canonical-bridge-server/src/shared/web3/web3.service.ts
@@ -12,21 +12,26 @@ import {
   CBRIDGE_ENDPOINT,
   CMC_API_ENDPOINT,
   CMC_API_KEY,
+  CMC_TOKEN_REQUEST_LIMIT,
   COINGECKO_ENDPOINT,
   DEBRIDGE_ENDPOINT,
-  STARGATE_ENDPOINT,
-  MESON_ENDPOINT,
   LLAMA_COINS_ENDPOINT,
-  CMC_TOKEN_REQUEST_LIMIT,
+  MAYAN_ENDPOINT,
+  MESON_ENDPOINT,
+  SIA_MAYAN_ENDPOINT,
   STARGATE_CHAIN_INFO,
+  STARGATE_ENDPOINT,
 } from '@/common/constants';
-import { values } from 'lodash';
+import { pick, values } from 'lodash';
 import {
   ICBridgeTransferConfig,
   IDeBridgeChain,
   IDeBridgeToken,
   IMesonTransferConfig,
   IStargateTokenList,
+  IMayanChain,
+  IMayanTransferConfig,
+  IMayanToken,
 } from '@bnb-chain/canonical-bridge-sdk';
 
 @Injectable()
@@ -138,6 +143,30 @@ export class Web3Service {
     } catch (e) {
       this.logger.log(`Failed to retrieve Meson API data at ${new Date().getTime()}, ${e}`);
       return [];
+    }
+  }
+
+  async getMayanConfigs(): Promise<IMayanTransferConfig> {
+    try {
+      const { data } = await this.httpService.axiosRef.get<{ chains: Array<IMayanChain> }>(
+        `${SIA_MAYAN_ENDPOINT}/init`,
+      );
+      const supportedChains = data.chains.filter(
+        (chain) => chain.originActive && chain.destinationActive,
+      );
+      const { data: tokens } = await this.httpService.axiosRef.get<
+        Record<string, Array<IMayanToken>>
+      >(`${MAYAN_ENDPOINT}/tokens?nonPortal=true`);
+      const supportedTokens = pick(
+        tokens,
+        supportedChains.map((c) => c.nameId),
+      );
+      return {
+        chains: supportedChains,
+        tokens: supportedTokens,
+      };
+    } catch (e) {
+      this.logger.log(`Failed to retrieve cBridge data at ${new Date().getTime()}, ${e.message}`);
     }
   }